requirements.txt
a2wsgi==1.10.10
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.10.0
apache-airflow==3.0.4
apache-airflow-core==3.0.4
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.27.5
apache-airflow-providers-smtp==2.2.0
apache-airflow-providers-standard==1.6.0
apache-airflow-task-sdk==1.0.4
argcomplete==3.1.4
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Automat==22.10.0
Babel==2.10.3
backports.strenum==1.3.1
bcrypt==3.2.2
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
cadwyn==5.4.4
certifi==2023.11.17
chardet==5.2.0
click==8.2.1
cloud-init==25.2
colorama==0.4.6
colorlog==6.9.0
command-not-found==0.3
ConfigArgParse==1.7.1
configobj==5.0.8
constantly==23.10.4
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==41.0.7
db==0.1.1
db-dtypes==1.4.3
dbus-python==1.3.2
decorator==5.1.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
distro-info==1.7+build1
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email_validator==2.2.0
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.115.5
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
fsspec==2025.7.0
gevent==25.5.1
geventhttpclient==2.3.4
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.7.0
incremental==22.10.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kafka==1.3.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy-object-proxy==1.11.0
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
locust==2.38.1
locust-cloud==1.26.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgpack==1.1.1
msgspec==0.19.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
pandas-gbq==0.29.2
pandas-stubs==2.3.0.250703
pathspec==0.12.1
pendulum==3.1.0
pipx==1.4.3
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.19.0
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
pydantic==2.11.7
pydantic-settings==2.5.2
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==8.4.1
python-apt==2.7.7+ubuntu5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==27.0.1
redis==6.4.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
retryhttp==1.3.3
rich==13.7.1
rich-argparse==1.7.1
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.0
rsa==4.9.1
scipy==1.15.3
sentry-sdk==2.35.0
service-identity==24.1.0
setproctitle==1.3.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.41.3
structlog==25.4.0
svcs==25.1.0
systemd-python==235
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
texttable==1.6.7
tomli==2.2.1
tqdm==4.67.1
Twisted==24.3.0
typer==0.16.1
types-pytz==2025.2.0.20250809
types-requests==2.32.4.20250809
types-setuptools==80.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-pro-client==8001
uc-micro-py==1.0.3
unattended-upgrades==0.1
universal_pathlib==0.2.6
urllib3==2.0.7
userpath==1.9.1
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0
zope.event==5.1.1
zope.interface==6.1
